Move mapping out of module
