+++ website/composables/index.ts
@@ -1,3 +1,5 @@
 export * from "./useDirectus";
 export * from "./useIngredientMultiplier";
 export * from "./useRecipeFormatter";
+export * from "./useImage";
+export * from "./useMapper";
