+++ website/modules/recipe/index.ts
@@ -1,9 +1,7 @@
-import type { ServerImage, ServerRecipePreview } from "common/types/serverRecipe";
 import MiniSearch from "minisearch";
 import { defineNuxtModule } from "nuxt/kit";
-import prand from "pure-rand";
-import { useDirectus, useRecipeFormatter } from "../../composables";
-import type { Image, RecipePreview } from "../../types/recipe";
+import { useDirectus, useMapper } from "../../composables";
+import type { RecipePreview } from "../../types/recipe";
 import { searchIndexSettings, type SearchIndexIndexed } from "../../types/searchIndex";
 import * as fs from "fs/promises";
 import * as crypto from "crypto";
@@ -26,7 +24,7 @@ export default defineNuxtModule({
   },
 });
 
-async function getAllRecipes() {
+async function getAllRecipes(): Promise<RecipePreview[]> {
   console.log(`Loading recipes from ${process.env.NUXT_BASE_URL}`);
   const client = useDirectus({
     url: process.env.NUXT_BASE_URL!,
@@ -47,16 +45,8 @@ async function getAllRecipes() {
     throw new Error(`No recipes were retrieved from ${process.env.NUXT_BASE_URL}`);
   }
 
-  return remoteRecipes.map((serverRecipe) => {
-    return {
-      title: serverRecipe.title,
-      tags: buildTagList(serverRecipe),
-      featuredTag: getRandomTag(serverRecipe),
-      totalDuration: getTotalDuration(serverRecipe),
-      coverImage: serverRecipe.coverImage ? mapImage(serverRecipe.coverImage) : undefined,
-      slug: serverRecipe.slug,
-    };
-  });
+  const mapper = useMapper();
+  return remoteRecipes.map(mapper.toRecipePreview);
 }
 
 function generateRecipeSearchIndex(recipes: RecipePreview[]) {
@@ -76,51 +66,4 @@ async function saveRecipeSearchIndex(index: string, nuxt: Nuxt) {
   // Store a hash of the index in config for cache busting
   const hash = crypto.createHash("md5").update(index).digest("hex");
   nuxt.options.appConfig.searchIndex.hash = hash;
-}
-
-function mapImage(serverImage: ServerImage): Image {
-  return {
-    id: serverImage.id,
-    title: serverImage.title,
-    width: serverImage.width,
-    height: serverImage.height,
-    modifyDate: serverImage.modified_on,
-  };
-}
-
-function getRandomTag(recipe: ServerRecipePreview) {
-  const tags = buildTagList(recipe);
-  if (tags.length === 0) {
-    return undefined;
-  }
-
-  // Is this randomness enough with incrementing integers?
-  const randomness = prand.xoroshiro128plus(recipe.id);
-  const [randomIndex] = prand.uniformIntDistribution(0, tags.length - 1, randomness);
-  console.log(randomIndex);
-  return tags[randomIndex];
-}
-
-const formatter = useRecipeFormatter();
-function getTotalDuration(recipe: ServerRecipePreview) {
-  return formatter.formatMinutesAsDuration(
-    (recipe.preparationDuration ?? 0) + (recipe.cookingDuration ?? 0) + (recipe.customDuration ?? 0),
-  );
-}
-
-function buildTagList(recipe: ServerRecipePreview): string[] {
-  const tags: string[] = [];
-  if (recipe.cuisine) {
-    tags.push(recipe.cuisine);
-  }
-  if (recipe.course) {
-    tags.push(recipe.course);
-  }
-  if (recipe.diets) {
-    tags.push(...recipe.diets);
-  }
-  if (recipe.main_ingredients) {
-    tags.push(...recipe.main_ingredients);
-  }
-  return tags.sort();
 }
